backend/requirements-django.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-environ==0.11.2
boto3==1.34.0

# Include existing Phase 1 dependencies
huggingface_hub>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.2.0

